Finish(?) re-arranging routes/navs, add minimal request list item
